successful demo of sharing memory between oss and user.
A 3-element array is mapped to shared memory by oss.
oss can read and modify the values.
The user process (run independently) can then read and modify the values.
oss can then see the changes that user has made.
This forms the skeleton for our clock and pid values which will be stored in
this 3-element array for the project.

To make this work, shmfunctions.h has been imported from the last project.
This collection of functions serves as a wrapper/abstraction for using ftok,
shmget, shmctl, shmat, and shmdt.  Slight modifications have been made so that
constants are used as parameters.